--- app/packages/rehype-momiji/parser.ts
@@ -1,6 +1,0 @@
-import { unified } from "unified";
-import rehypeParse from "rehype-parse";
-
-const parser = unified().use(rehypeParse, { fragment: true });
-
-export { parser };
